Loads model
